Restructure the modules so that they are in the gather, extract and analyze modules
